[feat]: IssueInfo 구현
이슈상세페이지에서 EditMode를 구분하도록 하였고, edit 모드가 아닐 때
보여줄 IssueInfo 컴포넌트를 구현하였습니다.
